requirements.txt
openai
requests
PyMuPDF
faiss-cpu
sentence-transformers
python-dotenv
google.generativeai